+++ src/extension.ts
@@ -33,6 +33,7 @@ import {
   updateAngularPackages,
   switchComponentFile,
   runNpmScript,
+  showSignalGraph,
 } from './commands';
 import {
   runNpmInstall,
@@ -117,6 +118,7 @@ export function activate(context: vscode.ExtensionContext) {
     ),
     vscode.commands.registerCommand('angular-cli-plus.runNpmScript', () => runNpmScript()),
     vscode.commands.registerCommand('angular-cli-plus.checkMemoryLeaks', () => checkMemoryLeaks()),
+    vscode.commands.registerCommand('angular-cli-plus.showSignalGraph', () => showSignalGraph()),
     vscode.commands.registerCommand('angular-cli-plus.npmInstall', () => runNpmInstall(false)),
     vscode.commands.registerCommand('angular-cli-plus.npmCleanInstall', () => runNpmInstall(true)),
     vscode.commands.registerCommand('angular-cli-plus.checkDependencies', async () => {